backend/tools_deployment/pneumonia-cloudrun/requirements.txt
flask
gunicorn
torch
torchvision
transformers
pillow